Detect and colorize out-of-sync releases across environments
